moved the gathering redux function to interactive common added a constant
