+++ src/datascience-ui/interactive-common/redux/reducers/transfer.ts
@@ -9,13 +9,14 @@ import {
 } from '../../../../client/datascience/interactive-common/interactiveWindowTypes';
 import { CssMessages } from '../../../../client/datascience/messages';
 import { ICell } from '../../../../client/datascience/types';
-import { extractInputText, getSelectedAndFocusedInfo, IMainState } from '../../mainState';
+import { extractInputText, getSelectedAndFocusedInfo, ICellViewModel, IMainState } from '../../mainState';
 import { isSyncingMessage, postActionToExtension } from '../helpers';
 import { Helpers } from './helpers';
 import {
     CommonActionType,
     CommonReducerArg,
     ICellAction,
+    IChangeGatherStatus,
     IEditCellAction,
     ILinkClickAction,
     ISendCommandAction,
@@ -325,6 +326,26 @@ export namespace Transfer {
             // As soon as an untrusted notebook is loaded, prompt the user to trust it
             postActionToExtension(arg, InteractiveWindowMessages.LaunchNotebookTrustPrompt);
         }
+        return arg.prevState;
+    }
+
+    export function gathering(arg: CommonReducerArg<CommonActionType, IChangeGatherStatus>): IMainState {
+        const index = arg.prevState.cellVMs.findIndex((c) => c.cell.id === arg.payload.data.cellId);
+        if (index >= 0) {
+            const cellVMs = [...arg.prevState.cellVMs];
+            const current = arg.prevState.cellVMs[index];
+            const newCell: ICellViewModel = {
+                ...current,
+                gathering: arg.payload.data.gathering
+            };
+            cellVMs[index] = newCell;
+
+            return {
+                ...arg.prevState,
+                cellVMs
+            };
+        }
+
         return arg.prevState;
     }
 }
